+++ bfs_nvlink/figure5_prio.sh
@@ -13,11 +13,18 @@ else
 fi
 
 if test -f "$outfile"; then
-    echo "Remove $outfile"
-    rm $outfile
+	echo "Do you want to regenerate performance data?"
+	read -r ifgenerate
+	if [ $ifgenerate == "yes" ] || [ $ifgenerate == "Yes" ] || [ $ifgenerate == "YES" ] || [ $ifgenerate == "1" ]; then
+    	echo "Remove $outfile"
+    	rm $outfile
+	fi
 fi
-echo "Generating performance data"
-./run_bfs_prio.sh $modequick >> $outfile
+
+if ! test -f "$outfile"; then 
+	echo "Generating performance data"
+	./run_bfs_prio.sh $modequick >> $outfile
+fi
 
 datasets=("soc-LiveJournal1" "hollywood_2009" "indochina_2004" "twitter" "road_usa" "osm-eur")
 output=`awk '{if($1 == "ave" && $2 == "time:") print $3","}' $outfile`
